requirements.txt
templar==1.3.1
invoke==0.9.0
python3-pika==0.9.14
flask==0.10.1

